+++ src/main/java/kr/hhplus/be/server/application/product/dto/ProductSearchRequest.java
@@ -2,10 +2,12 @@ package kr.hhplus.be.server.application.product.dto;
 
 import kr.hhplus.be.server.domain.product.dto.ProductSearchDto;
 
+import java.math.BigDecimal;
+
 public record ProductSearchRequest(
         String productName,
-        Integer minPrice,
-        Integer maxPrice
+        BigDecimal minPrice,
+        BigDecimal maxPrice
 ) {
 
     public ProductSearchDto toSearchDto() {
