+++ src/main/java/kr/hhplus/be/server/domain/point/dto/PointSearchResult.java
@@ -2,10 +2,12 @@ package kr.hhplus.be.server.domain.point.dto;
 
 import kr.hhplus.be.server.domain.point.Point;
 
+import java.math.BigDecimal;
+
 public record PointSearchResult(
         Long id,
         Long userId,
-        int balance
+        BigDecimal balance
 ) {
 
     public static PointSearchResult fromEntity(Point point) {
